Test start/stop flow of all protocols
