+++ bin/start_protocol.py
@@ -15,33 +15,10 @@
 # Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 import sys
-from types import SimpleNamespace
 
 import gevent
 
-from conpot import protocols
-from conpot.core import initialize_vfs
-from conpot.utils.greenlet import spawn_test_server, teardown_test_server
-
-
-def _init_test_server_by_name(name, port=0):
-    server_class = protocols.name_mapping[name]
-
-    template = {
-        "guardian_ast": "guardian_ast",
-        "IEC104": "IEC104",
-        "kamstrup_management": "kamstrup_382",
-        "kamstrup_meter": "kamstrup_382",
-    }.get(name, "default")
-
-    # Required by SNMP
-    class Args(SimpleNamespace):
-        mibcache = None
-
-    if name in ("ftp", "tftp"):
-        initialize_vfs()
-
-    return spawn_test_server(server_class, template, name, args=Args(), port=port)
+from conpot.utils.greenlet import init_test_server_by_name, teardown_test_server
 
 
 def main():
@@ -64,7 +41,7 @@ def main():
     port = ports.get(name, 0)
 
     print(f"Starting '{name}'...")
-    server, greenlet = _init_test_server_by_name(name, port=port)
+    server, greenlet = init_test_server_by_name(name, port=port)
 
     try:
         gevent.wait()
